test dynamic resolve again

+++ Dockerfile
@@ -8,7 +8,14 @@ echo "deb https://patrick-zippenfenig.github.io/ecCodes-ubuntu/ jammy main" > /e
 wget https://apache.jfrog.io/artifactory/arrow/ubuntu/apache-arrow-apt-source-latest-jammy.deb
 apt install -y -V ./apache-arrow-apt-source-latest-jammy.deb
 apt update
-apt-cache depends libparquet-glib-dev
-apt install -y tzdata libnetcdf19 libeccodes0 bzip2 curl gir1.2-parquet-24.0
+echo "apt-cache depends"
+apt-cache depends libparquet-glib-dev | awk '/Depends: gir1\.2-parquet-/{print $2; exit}'
+GIR_PARQUET_PKG="$(apt-cache depends libparquet-glib-dev | awk '/Depends: gir1\.2-parquet-/{print $2; exit}')"
+echo "ENV: $GIR_PARQUET_PKG"
+if [[ -z "$GIR_PARQUET_PKG" ]]; then
+	echo "Unable to determine gir1.2-parquet package from libparquet-glib-dev dependencies" >&2
+	exit 1
+fi
+apt install -y tzdata libnetcdf19 libeccodes0 bzip2 "$GIR_PARQUET_PKG" curl
 rm -rf /var/lib/apt/lists/*
 SHELL
